refactor: Break apart Suggestion.tsx
