+++ src/components/Landing/MainForm/SpecInput/Suggestion/index.ts
@@ -1,0 +1,4 @@
+import Suggestion, { SuggestionProps } from "./Suggestion";
+
+export default Suggestion;
+export type { SuggestionProps };
